// frontend/src/app/modules/management/interfaces/IInternalProgram.ts
export interface IINternalProgram {
    creationDate: string;
    internalProgramGuid: string;
    internalProgramName: string;
    internalProgramDescription: string;
    internalProgramReviewCycle: string;
    internalProgramActiveDays: number;
    isInternalProgramCategoryExists: boolean;
    modificationDate: string;
    isActive:boolean;
}