// frontend/src/app/modules/candidate/interfaces/IBasicDetail.ts
export interface IBasicDetail {
    id: string;
    employeeCode: string;
    firstName: string;
    lastName: string;
    email: string;
    phoneNumber: string;
    emergencyContactNo: string;
    pronounGuid: string;
    pronoun: string;
    businessUnitGuid: string;
    businessUnitName: string;
    businessUnitLogoName: string;
    country: string;
    state: string;
    city: string;
    designation: string;
    birthDate: string;
    about: string;
    profilePhotoName: string;
    joiningDate: string;
    skypeId: string;
    linkedInUrl: string;
}